fixed a bug that causes the memory leak in the validation function Note: when create a Variable and don't do anything (no backward), there may be a memory leak. It seems that the PyTorch will not automatically use GC for the Variables.

+++ layers.py
@@ -70,7 +70,7 @@ class HM_LSTMCell(Module):
         #     c_new = i * g
         #     h_new = o * Func.tanh(c_new)
 
-        z_new = bound(z_hat).float()
+        z_new = bound()(z_hat)
 
         return h_new, c_new, z_new
